authorizations done

+++ app/Http/Controllers/AuthorizationsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\AuthorizationsRequest;
+use App\Models\Audit;
 use App\Models\Authorization;
 use App\Models\Role;
 use App\Models\User;
@@ -17,16 +18,17 @@ class AuthorizationsController extends Controller
      */
     public function index(AuthorizationsRequest $request)
     {
+        $this->authorize('viewAny', Authorization::class);
+
         $request = $request->validated();
         extract($request);
 
         try {
-            $authorizations = Authorization::query()
+            $authorizations = Authorization::select('id', 'user_id', 'role_id', 'status', 'created_at')
                 ->with([
                     'user' => fn($user) => $user->select('id', 'name'),
                     'role' => fn($role) => $role->select('id', 'role_name')
                 ])
-                ->select('id', 'user_id', 'role_id', 'status', 'created_at')
                 ->when($search, function($query, $search){
                     $query->whereHas('user', function($user) use($search){
                         $user->where('name', 'like', "%{$search}%");
@@ -50,9 +52,10 @@ class AuthorizationsController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Authorization::class);
+
         try {
-            $users = User::query()
-                ->select('id', 'name')
+            $users = User::select('id', 'name')
                 ->whereNotIn('id', Authorization::pluck('user_id'))
                 ->orderBy('name')
                 ->get();
@@ -73,6 +76,8 @@ class AuthorizationsController extends Controller
      */
     public function store(AuthorizationsRequest $request)
     {
+        $this->authorize('create', Authorization::class);
+
         try {
             $create = Authorization::create($request->validated());
             return redirect()->route('authorizations.show', $create->id)->with('success', __('Saved'));
@@ -89,7 +94,14 @@ class AuthorizationsController extends Controller
      */
     public function show($id)
     {
-        return view('pages.authorizations.show');
+        $this->authorize('view', Authorization::class);
+
+        try {
+            $authorization = Authorization::select('id', 'user_id', 'role_id', 'status', 'created_at')->find($id);
+            return view('pages.authorizations.show', compact('authorization'));
+        } catch (\Throwable $th) {
+            Log::channel('catch')->info($th);
+        }
     }
 
     /**
@@ -100,7 +112,15 @@ class AuthorizationsController extends Controller
      */
     public function edit($id)
     {
-        //
+        $this->authorize('update', Authorization::class);
+
+        try {
+            $authorization = Authorization::select('id', 'user_id', 'role_id', 'status')->find($id);
+            $roles = Role::select('id', 'role_name')->orderBy('role_name')->get();
+            return view('pages.authorizations.edit', compact('authorization', 'roles'));
+        } catch (\Throwable $th) {
+            Log::channel('catch')->info($th);
+        }
     }
 
     /**
@@ -112,7 +132,14 @@ class AuthorizationsController extends Controller
      */
     public function update(AuthorizationsRequest $request, $id)
     {
-        //
+        $this->authorize('update', Authorization::class);
+
+        try {
+            Authorization::find($id)->update($request->validated());
+            return redirect()->route('authorizations.show', $id)->with('success', __('Updated'));
+        } catch (\Throwable $th) {
+            Log::channel('catch')->info($th);
+        }
     }
 
     /**
@@ -123,9 +150,45 @@ class AuthorizationsController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', Authorization::class);
+
         try {
             Authorization::destroy($id);
             return redirect()->back()->with('success', __('Deleted'));
+        } catch (\Throwable $th) {
+            Log::channel('catch')->info($th);
+        }
+    }
+
+        /**
+     * Display the audits of specified resource.
+     *
+     * @param  int  $id
+     * @param  \App\Http\Requests\AuthorizationsRequest  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function audits(AuthorizationsRequest $request, $id)
+    {
+        $this->authorize('view', Authorization::class);
+        
+        $request = $request->validated();
+        extract($request);
+
+        try {
+            $audits = Audit::select('id', 'user_id', 'event', 'old_values', 'new_values')
+                ->addSelect('user_agent', 'ip_address', 'created_at')
+                ->whereAuditableType('App\\Models\\Authorization')
+                ->whereAuditableId($id)
+                ->when($search, function($query, $search){
+                    $query->whereHas('user', function($user) use($search){
+                        $user->where('name', 'like', "%{$search}%");
+                    });
+                })
+                ->orderByDesc('id')
+                ->paginate(10)
+                ->appends($request);
+
+            return view('pages.authorizations.audits', compact('id', 'audits'));
         } catch (\Throwable $th) {
             Log::channel('catch')->info($th);
         }
